+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Change Log
 * Allow to start Black Mamba even in untested version of Pythonista, just
   warn the user
 * Init messages are colored (orange=warn, red=error)
+* All print messages replaced with ``log.info`` (``.error``, ``.warn``)
+* ``bm.log.level`` allows to set logging level (default ``INFO``)
  
 
 ``0.0.15`` (2017-08-28)